Implemented martijn's improvements to glabals diffing from the other branch.
Slight improvement: whitespace differences are also ignored along with comments.
Also comments are stripped more robustltly than with .split('#').
Both of these are achieved with the stdlit tokenize module (which is
exactly for this sort of purpose).

As in the other branch's implementation, differences are only reported
if both the value and the expression that produced it (ignoring comments
and whitespace) differ. This is important because some types of objects
always give False on equality checks, so that produces false positives.

Fixed bug in dict_diff that improves its comparison of nnarrays.

Made ctrl-c kill the program

Made groups open on double click (and go to the appropriate tab)