Fixes bug with the implementation of #36.
Expansion of globals (to calculated number of shots to compile) is now deferred until after the expansion groups have been updated.
This fixes an exception that was raised when changing a global part of an automatic zip group (one created because a second global referenced the first) from a list to a single value.