Add the concept of a sequence_index to runmanager.
The sequence index is an incrementing integer starting from zero each day for
each experiment script. It is stored on disk in
<experiment_shot_storage>/<script_name>/.next_sequence_index and accessed using
zlock so that it may be incremented atomically.

The sequence index may be used in the output folder format, which defaults to:
output_folder_format = %%Y/%%m/%%d/{sequence_index:05d}, resulting in output
folders like:
<shared_drive>/Experiments/<experiment_name>/<script_name>/2019/03/14/00006

The format string will be fed first to strftime, then a call to .format()
with sequence_index and sequence_timestamp kwargs, where sequence_timestamp is
the time in format '%Y%m%dT%H%M%S', which is the existing format used in
sequence_ids.

Allow the shot filename to also have its format configured, such as:
filename_prefix_format = '%%Y-%%m-%%d_{sequence_index:05d}_{script_basename}',
resulting in shot filenames like:
2019-03-14_00006_myscript_1.h5

This is similarly passed through strftime and then .format with sequence_index,
sequence_timestamp, and script_basename kwargs.

The pre-set output_folder_format if newly installed labconfigs will be:
%%Y/%%m/%%d/{sequence_index:05d}
which will also be the default if not present in the labconfig file.

The pre-set filename_prefix_format in newly installed labconfigs will be:
%%Y-%%m-%%d_{sequence_index:05d}_{script_basename}
but if not present in labconfig files, {sequence_timestamp}_{script_basename}
will be used to preserve the present behaviour of using the sequence id.

For compatibility with lyse, the sequence_id is still saved to the h5 file, but
other attributes are saved too. These are:

'script_basename': the basename of the script, i.e. myscript
'sequence_date': a string of the date only, i.e. 2019-03-14
'sequence_index': the sequence_index as an integer,
'sequence_id': the sequence_id, i.e. 20190314183955_myscript

The idea is that script_basename, sequence_date and sequence_id become an
alternative to sequence_id as a way of uniquely identifying a shot. lyse
already has a labconfig setting where it uses the sequence_index if present to
index the dataframe (added for the  sake of Spielman fork shots which already
have sequence_index), but it has the issue that shots cannot be fully uniquely
determined by that alone since the date and script name are needed too. So
the date and script name are added to the attrs for this purpose, but sequence_id
is retained.

Refactored runmanager functions to achieve this. There is now a fairly monolithic function
new_sequence_details(), which returns the attrs, folder and filepath prefix for
the given labconfig settings and current time. It's all in one function because
the different pieces all use timestamps, and so to avoid races it either has to be
a single function call or the caller would need to carry around a datetime object.
Since it is not too complicated I made it a single function call, though the callers
are often ignoring one or more of its return values.

Removed generate_sequence_id, since the functions that accept a sequence_id
now accept the sequence_attrs dict instead (one entry of which is the sequence_id).
This and the modifying of other function call signatures is a backward incompatible
change in the sense of the API that runmanager exposes, but is backward-compatible
with how other labscript programs use the runmanager API, i.e. BLACS. So I think
this is acceptable - code using the runmanager API directly may need to change.
We could do a major version bump to indicate this.